Prevent any keystroke to delete the figure Only the deletion and the backspace key will be able to remove the selected figure

+++ lib/user_interaction/keyboard_controller.dart
@@ -10,9 +10,16 @@ class KeyboardController extends GraphObject with EndlinePropagator {
   }
 
   void onKeyDown(KeyDownEvent key) {
+    if (!isADeletionKey(key)) return;
     final objectDeletion = interactionReference.deleteInterface;
     if (objectDeletion == null) return;
     objectDeletion.delete();
     setState(this);
+  }
+
+  bool isADeletionKey(KeyDownEvent key) {
+    if (key.logicalKey == LogicalKeyboardKey.backspace) return true;
+    if (key.logicalKey == LogicalKeyboardKey.delete) return true;
+    return false;
   }
 }
